Dissolved class FlightMgr
- integrated methods of Planes friend class into Planes
